% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_col.R
\name{draw_col}
\alias{draw_col}
\title{Draw column chart}
\usage{
draw_col(df)
}
\arguments{
\item{df}{A data frame.}
}
\value{
A plot.
}
\description{
Draw column chart
}
\examples{
\dontrun{
draw_col(data())
}
}
