% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_map.R
\name{draw_map}
\alias{draw_map}
\title{Draw Japan prefecture map}
\usage{
draw_map(vec)
}
\arguments{
\item{vec}{A numeric vector of length 47 in order of jiscode.}
}
\value{
A plot.
}
\description{
Draw Japan prefecture map
}
\examples{
\dontrun{
draw_map(data())
}
}
