% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findItem.R
\name{findItemServer}
\alias{findItemServer}
\title{findItem module Server}
\usage{
findItemServer(id)
}
\arguments{
\item{id}{A character vector of length 1.}
}
\value{
A module server.
}
\description{
findItem module Server
}
\examples{
\dontrun{
findItemServer("city")
}
}
