% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findItem.R
\name{findItemUI}
\alias{findItemUI}
\title{findItem module UI}
\usage{
findItemUI(id)
}
\arguments{
\item{id}{A character vector of length 1.}
}
\value{
A module UI.
}
\description{
findItem module UI
}
\examples{
\dontrun{
findItemUI("city")
}
}
