% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{mainApp}
\alias{mainApp}
\title{Shiny app to show Japanese household expenditures by prefecture}
\usage{
mainApp()
}
\value{
A Shiny app.
}
\description{
Shiny app to show Japanese household expenditures by prefecture
}
\examples{
\dontrun{
mainApp()
}
}
