% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showMap.R
\name{showMapServer}
\alias{showMapServer}
\title{showMap module Server}
\usage{
showMapServer(id)
}
\arguments{
\item{id}{A character vector of length 1.}
}
\value{
A module server.
}
\description{
showMap module Server
}
\examples{
\dontrun{
showMapServer("item")
}
}
