% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showMap.R
\name{showMapUI}
\alias{showMapUI}
\title{showMap module UI}
\usage{
showMapUI(id)
}
\arguments{
\item{id}{A character vector of length 1.}
}
\value{
A module UI.
}
\description{
showMap module UI
}
\examples{
\dontrun{
showMapUI("item")
}
}
